@@ -7,12 +7,13 @@
     "build": "webpack --env.production"
   },
   "dependencies": {
-    "@babel/core": "^7.13.14",
-    "@babel/preset-env": "^7.13.12",
+    "@babel/core": "^7.13.15",
+    "@babel/preset-env": "^7.13.15",
     "@babel/preset-react": "^7.13.13",
     "@react-pdf/renderer": "^2.0.0",
     "babel-loader": "^8.2.2",
     "brfs": "^2.0.2",
+    "buffer": "^6.0.3",
     "canvg": "^3.0.7",
     "copy-webpack-plugin": "^8.1.1",
     "deep-equal": "^2.0.5",
@@ -22,13 +23,15 @@
     "html2canvas": "^1.0.0-rc.7",
     "lodash.omit": "^4.5.0",
     "prettier": "^2.2.1",
+    "process": "^0.11.10",
     "queue": "^6.0.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-vis": "^1.11.7",
     "recharts": "^2.0.9",
     "regenerator-runtime": "^0.13.7",
-    "webpack": "^5.30.0",
+    "util": "^0.12.3",
+    "webpack": "^5.31.0",
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2"
   },
